requirements.txt
flask==3.0.0
boto3==1.34.0
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
pillow==9.5.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-talisman==1.1.0
